crewai-agent/requirements.txt
khaos-agent>=1.0.0
crewai>=1.6.1
langchain-openai>=0.2
openai>=1.0
